backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.10.0
    # via
    #   channels
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.10.2
    # via daphne
automat==25.4.16
    # via twisted
awsebcli==3.25.3
    # via -r requirements.in
bcrypt==5.0.0
    # via paramiko
black==25.11.0
    # via -r requirements.in
blessed==1.23.0
    # via awsebcli
boto3==1.40.72
    # via -r requirements.in
botocore==1.40.72
    # via
    #   awsebcli
    #   boto3
    #   s3transfer
cement==2.10.14
    # via awsebcli
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via
    #   cryptography
    #   pynacl
channels==4.1.0
    # via -r requirements.in
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via black
colorama==0.4.6
    # via awsebcli
constantly==23.10.4
    # via twisted
coverage==7.11.3
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==46.0.3
    # via
    #   autobahn
    #   paramiko
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
decorator==5.2.1
    # via fabric
deprecated==1.3.1
    # via fabric
django==5.2.8
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-admin-interface==0.31.0
    # via -r requirements.in
django-colorfield==0.14.0
    # via
    #   -r requirements.in
    #   django-admin-interface
django-cors-headers==4.9.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
execnet==2.1.2
    # via pytest-xdist
fabric==3.2.2
    # via awsebcli
factory-boy==3.3.3
    # via -r requirements.in
faker==38.0.0
    # via
    #   -r requirements.in
    #   factory-boy
flake8==7.3.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
iniconfig==2.3.0
    # via pytest
invoke==2.2.1
    # via
    #   fabric
    #   paramiko
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
packaging==24.2
    # via
    #   awsebcli
    #   black
    #   gunicorn
    #   pytest
paramiko==4.0.0
    # via fabric
pathspec==0.12.1
    # via
    #   awsebcli
    #   black
pillow==12.0.0
    # via
    #   -r requirements.in
    #   django-colorfield
platformdirs==4.5.0
    # via black
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycodestyle==2.14.0
    # via flake8
pycparser==2.23
    # via cffi
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pymysql==1.1.2
    # via -r requirements.in
pynacl==1.6.1
    # via paramiko
pyopenssl==25.3.0
    # via twisted
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-django
    #   pytest-xdist
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-cov==7.0.0
    # via -r requirements.in
pytest-django==4.11.1
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   awsebcli
    #   botocore
python-dotenv==1.2.1
    # via -r requirements.in
python-slugify==8.0.4
    # via
    #   -r requirements.in
    #   django-admin-interface
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via awsebcli
requests==2.32.5
    # via
    #   -r requirements.in
    #   awsebcli
s3transfer==0.14.0
    # via boto3
semantic-version==2.10.0
    # via awsebcli
service-identity==24.2.0
    # via twisted
setuptools==80.9.0
    # via
    #   awsebcli
    #   incremental
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
termcolor==2.5.0
    # via awsebcli
text-unidecode==1.3
    # via python-slugify
twisted==25.5.0
    # via daphne
txaio==25.9.2
    # via autobahn
typing-extensions==4.15.0
    # via
    #   pyopenssl
    #   pytest-asyncio
    #   twisted
tzdata==2025.2
    # via faker
urllib3==1.26.20
    # via
    #   awsebcli
    #   botocore
    #   requests
wcwidth==0.2.14
    # via
    #   awsebcli
    #   blessed
whitenoise==6.11.0
    # via -r requirements.in
wrapt==2.0.1
    # via deprecated
zope-interface==8.1
    # via twisted
